requirements.txt
absl-py==1.0.0
astunparse==1.6.3
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cv==1.0.0
Flask==2.0.2
Flask-Cors==3.0.10
flatbuffers==2.0
gast==0.4.0
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.22.1
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.4.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
